[DateTimePicker] Sort out tests
